refactor: rename many fields and files
