+++ integrations/maya/src/main/res/layout/content_conversion_preview_maya.xml
@@ -188,7 +188,7 @@
         app:layout_constraintBottom_toBottomOf="@+id/purchase_label" />
 
     <androidx.constraintlayout.widget.ConstraintLayout
-        android:id="@+id/coinbase_fee_info_container"
+        android:id="@+id/maya_fee_info_container"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:orientation="horizontal"
@@ -197,7 +197,7 @@
         app:layout_constraintEnd_toEndOf="parent">
 
         <TextView
-            android:id="@+id/coinbase_fee_label"
+            android:id="@+id/maya_fee_label"
             style="@style/Caption.Secondary"
             android:layout_width="wrap_content"
             android:layout_height="46dp"
@@ -214,28 +214,28 @@
             android:layout_height="wrap_content"
             android:padding="7dp"
             android:src="@drawable/ic_dash_info_blue_meduim"
-            app:layout_constraintBottom_toBottomOf="@+id/coinbase_fee_label"
-            app:layout_constraintStart_toEndOf="@id/coinbase_fee_label"
-            app:layout_constraintTop_toTopOf="@id/coinbase_fee_label" />
+            app:layout_constraintBottom_toBottomOf="@+id/maya_fee_label"
+            app:layout_constraintStart_toEndOf="@id/maya_fee_label"
+            app:layout_constraintTop_toTopOf="@id/maya_fee_label" />
 
         <TextView
-            android:id="@+id/coinbase_fee_amount"
+            android:id="@+id/maya_fee_amount"
             style="@style/Caption"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            app:layout_constraintBottom_toBottomOf="@id/coinbase_fee_label"
+            app:layout_constraintBottom_toBottomOf="@id/maya_fee_label"
             app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintTop_toTopOf="@id/coinbase_fee_label"
+            app:layout_constraintTop_toTopOf="@id/maya_fee_label"
             tools:text="$ 1.00" />
 
     </androidx.constraintlayout.widget.ConstraintLayout>
 
     <View
-        android:id="@+id/coinbase_fee_separator"
+        android:id="@+id/maya_fee_separator"
         android:layout_width="match_parent"
         android:layout_height="1dp"
         android:background="@color/divider_color"
-        app:layout_constraintBottom_toBottomOf="@+id/coinbase_fee_info_container" />
+        app:layout_constraintBottom_toBottomOf="@+id/maya_fee_info_container" />
 
     <TextView
         android:id="@+id/total_label"
@@ -246,7 +246,7 @@
         android:text="@string/total"
         android:theme="@style/ListViewButtonBackground"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/coinbase_fee_info_container" />
+        app:layout_constraintTop_toBottomOf="@id/maya_fee_info_container" />
 
     <TextView
         android:id="@+id/total_amount"
