+++ integrations/maya/src/main/res/layout/dialog_maya_error.xml
@@ -60,7 +60,7 @@
 
     <androidx.appcompat.widget.AppCompatButton
         android:id="@+id/coinbase_dialog_negative_button"
-        style="@style/Button.Tertiary.Blue"
+        style="@style/Button.Primary.Blue"
         android:layout_width="0dp"
         android:layout_height="match_parent"
         android:layout_marginTop="5dp"
