+++ integrations/maya/src/main/res/layout/dialog_maya_fee_info.xml
@@ -36,7 +36,7 @@
             app:layout_constraintTop_toBottomOf="@id/maya_fee_info_icon"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
-            android:text="@string/fees_in_crypto_purchase"
+            android:text="@string/maya_fees_in_crypto_purchase"
             style="@style/Subtitle1"
             android:textSize="16sp"
             android:layout_marginTop="32dp"
@@ -52,7 +52,7 @@
             style="@style/Caption.Secondary"
             android:layout_marginTop="5dp"
             android:textSize="12sp"
-            android:text="@string/coinbase_fee_info_msg_content"
+            android:text="@string/maya_fee_info_msg_content"
             android:layout_marginHorizontal="20dp"
             android:paddingVertical="10dp"
             android:gravity="center"
@@ -80,8 +80,7 @@
             android:layout_marginLeft="20dp"
             android:layout_marginRight="20dp"
             android:layout_marginTop="36dp"
-            android:background="@drawable/round_corner_grey_border"
-            style="@style/Button.Secondary"
+            style="@style/Button.Primary.Blue"
             android:text="@string/button_close"
             android:textAllCaps="false"
             />
