+++ integrations/maya/src/main/res/layout/dialog_maya_result.xml
